requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── AI / LLM ───────────────────────────────────────────────────────────────────
langchain==0.2.16
langchain-core==0.2.38
langchain-groq==0.1.9
groq==0.12.0
httpx==0.27.2
# ── Document Extraction ────────────────────────────────────────────────────────
PyMuPDF==1.24.12          # PDF extraction (fitz)
python-docx==1.1.2        # DOCX extraction

# ── OCR ────────────────────────────────────────────────────────────────────────
easyocr==1.7.2            # Primary OCR (deep learning)
pytesseract==0.3.13       # Fallback OCR
Pillow==11.0.0            # Image processing

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.9.2
numpy==1.26.4


google-cloud-vision